move SAS examples to their own class
