+++ blobs/howto/dotnet/dotnet-v12/dotnet-v12.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Azure.Identity" Version="1.3.0" />
     <PackageReference Include="Azure.Storage.Blobs" Version="12.7.0" />
     <PackageReference Include="Azure.Storage.Files.DataLake" Version="12.5.0" />
     <PackageReference Include="Azure.Storage.Queues" Version="12.5.0" />
